Separate protocols Highlightable and Enableable from UILabel extension.
These are used in files unrelated to UILabel. Also, changed the parameter names in the corresponding extensions from `label` to more generic names.
