+++ Sources/UIKitExtensions/UILabel.swift
@@ -1,30 +1,6 @@
 import UIKit
 
-public protocol Highlightable: class {
-    var isHighlighted: Bool { get set }
-}
-
-public protocol Enableable: class {
-    var isEnabled: Bool { get set }
-}
-
 extension UILabel: Highlightable, Enableable {}
-
-public extension ConfigurationSet where Base: Highlightable {
-    func isHighlighted(_ highlighted: Bool) -> ConfigurationSet<Base> {
-        return set { (label: Base) in
-            label.isHighlighted = highlighted
-        }
-    }
-}
-
-public extension ConfigurationSet where Base: Enableable {
-    func isEnabled(_ enabled: Bool) -> ConfigurationSet<Base> {
-        return set { (label: Base) in
-            label.isEnabled = enabled
-        }
-    }
-}
 
 public extension ConfigurationSet where Base: UILabel {
     func text(_ text: String) -> ConfigurationSet<Base> {
